add canvas module
