+++ index.html
@@ -59,6 +59,6 @@
         />
       </details>
     </main>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
